+++ src/components/Overview.jsx
@@ -40,38 +40,60 @@ export const Overview = ({ setTabIndex }) => {
       </Grid>
       <Grid item xs={10}>
         <Paper>
-          <Container>
-            <Box color='tertiary.main'>
-              <Typography variant='h5'>Goals</Typography>
+          <Box p={2}>
+            <Box display="flex" flexDirection="column" alignItems="center">
+              <Box color='tertiary.main'>
+                <Typography variant='h5'>Goals</Typography>
+              </Box>
+              <Box mb={2}>
+                <Typography fontFamily='Roboto Slab' variant='h3'>${UserCtx.user.total_saved}</Typography>
+              </Box>
             </Box>
-            <Box mb={2}>
-              <Typography fontFamily='Roboto Slab' variant='h3'>${UserCtx.user.total_saved}</Typography>
-            </Box>
-            <GoalLabels />
-            {goals.map(goal => <Goal key={goal.id} goal={goal} /> )}
-            <Box m={2} textAlign='center'>
-              <Button onClick={() => setTabIndex(1)}>
-                <Typography variant='overline'>See All</Typography>
-              </Button>
-            </Box>
-          </Container>
+            {goals.length > 0
+              ? (
+                <Box>
+                  <GoalLabels />
+                  {goals.map(goal => <Goal key={goal.id} goal={goal} /> )}
+                  <Box m={2} textAlign='center'>
+                    <Button onClick={() => setTabIndex(1)}>
+                      <Typography variant='overline'>See All</Typography>
+                    </Button>
+                  </Box>
+                </Box>
+              ) : (
+                <Box color="#C0C0C0" fontSize={14} textAlign="center">
+                  You haven't created any goals yet.
+                </Box>
+              )
+            }
+          </Box>
         </Paper>
       </Grid>
       <Grid item xs={10}>
         <Paper>
-            <Container>
-              <Box mb={2} color='secondary.main'>
-                <Typography>Transactions</Typography>
-              </Box>
-              <TransactionLabels />
-              {transactions.map((trx, idx) => <Transaction key={idx} trx={trx} /> )}
-              <Box m={2} textAlign='center'>
-              <Button onClick={() => setTabIndex(2)}>
-                <Typography variant='overline'>See All</Typography>
-              </Button>
+          <Box p={2}>
+            <Box mb={2} color='secondary.main'>
+              <Typography align="center" variant="h5">Transactions</Typography>
             </Box>
-            </Container>
-          </Paper>
+            {transactions.length > 0
+              ? (
+                <Box>
+                  <TransactionLabels />
+                  {transactions.map((trx, idx) => <Transaction key={idx} trx={trx} /> )}
+                  <Box m={2} textAlign='center'>
+                    <Button onClick={() => setTabIndex(2)}>
+                      <Typography variant='overline'>See All</Typography>
+                    </Button>
+                  </Box>
+                </Box>
+              ) : (
+                <Box color="#C0C0C0" fontSize={14} textAlign="center">
+                  You haven't created any transactions yet.
+                </Box>
+              )
+            }
+          </Box>
+        </Paper>
       </Grid>
     </Grid>
   );
